+++ CHANGELOG.md
@@ -16,12 +16,17 @@ Our versioning strategy is as follows:
 
 * `[nextjs]` `[App Router]` Form component throws a Server component error ([#333](https://github.com/Sitecore/content-sdk/pull/333))
 * `[nextjs]` `[react]` Fix fields becoming uneditable in Pages when running Editing Host in dev mode ([#339](https://github.com/Sitecore/content-sdk/pull/339))
+* `[nextjs]` Add "use client" directive to import-map.ts for React hooks compatibility ([#326](https://github.com/Sitecore/content-sdk/pull/326))
 
 ## 1.3.2
 
 ### ✨ Chores
 
 * Apply caret (`^`) verisoning to content-sdk packages, ensuring the latest patch versions are used by them.
+
+### 🐛 Bug Fixes
+
+* `[nextjs]` Add "use client" directive to import-map.ts for React hooks compatibility ([#326](https://github.com/Sitecore/content-sdk/pull/326))
 
 ## 1.3.1